Clear every contract xfail: both tables are now empty
The suite shipped 22 known violations across 8 algorithms and 1 splitter, with
a comment saying an empty table was the goal. It is empty.

Three of them were not algorithm bugs.

The contract fixture put NaN in the regression target. make_data punched holes
into X and then computed y from X[:, 1], so the target inherited them. The
author had protected column 0 so every row kept a usable value, but the target
reads column 1. XGBoostRegressor and CatBoostRegressor were recorded as
violating a contract they honour: missing_values is about features, and no
supervised learner trains on missing labels. The target is built before the
holes now, with an assertion so it cannot regress silently.

AgglomerativeClusterer.predict told the caller to refit with store_data=True, a
constructor argument that has never existed, and then raised
NotImplementedError unconditionally -- so the branch was dead and the advice
impossible. Agglomerative clustering genuinely has no out-of-sample rule, but
DBSCANClusterer already answers the same question by assigning to its nearest
core sample, so predict now assigns to the nearest of the centroids recorded at
fit time. It reproduces the fitted labels exactly on the training data.
fit_predict remains the right call when those labels are what you want.

KMeansClusterer declared missing_values it does not implement: the squared
distance to a centroid is undefined with NaN and nothing here imputes.
Declaring it was worse than declaring nothing, because the workflow layer
routes incomplete data on this capability. Dropped rather than faked.

NaiveBayesMultinomialClassifier declared plain numeric while correctly
rejecting negative features, as scikit-learn's MultinomialNB does -- the
multinomial likelihood is defined over counts, not the real line. The
capability vocabulary had no way to say so, so non_negative is added and the
sweep feeds it counts. Also removed a duplicate KNOWN_CAPABILITIES definition
that silently shadowed the first; they were identical, but an edit to the
earlier one would have had no effect.

DBSCANClusterer raised IndexError when no core samples were found, which is a
legitimate outcome on sparse data: np.array([]) on an empty set is float64, and
a float array cannot index. dtype=int, and the run returns all-noise labels.

StratifiedGroupKFold produced empty folds. It picked each group's fold by the
majority-class counter alone, so the classes filled folds independently from
the left -- three class-0 groups and three class-1 groups over five folds all
landed in folds 0 to 2, giving test sizes of [20, 20, 20, 0, 0] where
GroupKFold managed [20, 10, 10, 10, 10] on the same data. A fold with no test
rows scores nothing while still counting as a fold. Fold size is the tiebreak
now; groups stay disjoint.

Suite: 3,308 passed, 127 skipped, 0 xfailed.

+++ tuiml/algorithms/clustering/dbscan.py
@@ -294,7 +294,11 @@ class DBSCANClusterer(Clusterer):
                 cluster_id += 1
 
         self.labels_ = labels
-        self.core_sample_indices_ = np.array(sorted(core_samples))
+        # dtype=int explicitly: np.array([]) on an empty set is float64, and a
+        # float array cannot index, so a run that finds no core samples -- every
+        # point noise, which is a legitimate DBSCAN outcome on sparse data --
+        # raised IndexError on the next line instead of returning all -1.
+        self.core_sample_indices_ = np.array(sorted(core_samples), dtype=int)
         self.components_ = X[self.core_sample_indices_].copy()
 
         # Count clusters (excluding noise labeled as -1)
